% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-helpers.R
\name{lenv}
\alias{lenv}
\title{List environment variables with additional info}
\usage{
lenv(env = .GlobalEnv)
}
\arguments{
\item{env}{Environment to list}
}
\value{
data frame with the environment
}
\description{
List environment variables with additional info
}
